+++ src/components/HomepageFeatures/index.tsx
@@ -14,10 +14,12 @@ type FeatureItem = {
 const FeatureList: FeatureItem[] = [
   {
     title: "Getting Started",
-    Svg : require("@site/static/img/extension.svg").default,
+    Svg: require("@site/static/img/extension.svg").default,
     description: (
       <>
-        Discover the fundamentals of Rarimo Protocol, its architecture, applications, and advantages. Get familiar with cross-chain identity and bridging capabilities.
+        Discover the fundamentals of Rarimo Protocol, its architecture,
+        applications, and advantages. Get familiar with cross-chain identity and
+        bridging capabilities.
       </>
     ),
     url: "./overview",
@@ -27,7 +29,8 @@ const FeatureList: FeatureItem[] = [
     Svg: require("@site/static/img/arrow_top_right.svg").default,
     description: (
       <>
-        Learn the ropes of identity protocol, zero-knowledge proofs, cross-chain token bridging, and other tech by building real DApps from scratch.
+        Learn the ropes of identity protocol, zero-knowledge proofs, cross-chain
+        token bridging, and other tech by building real DApps from scratch.
       </>
     ),
     url: "./tutorials",
@@ -37,7 +40,8 @@ const FeatureList: FeatureItem[] = [
     Svg: require("@site/static/img/flag.svg").default,
     description: (
       <>
-        Integrate Rarimo products into your DApps, solve typical tasks, and troubleshoot common problems by following step-by-step instructions.
+        Integrate Rarimo products into your DApps, solve typical tasks, and
+        troubleshoot common problems by following step-by-step instructions.
       </>
     ),
     url: "./how-to-guides",
@@ -47,7 +51,8 @@ const FeatureList: FeatureItem[] = [
     Svg: require("@site/static/img/arrow_top_right.svg").default,
     description: (
       <>
-        Look up Rarimo's APIs, JavaScript libraries, and smart contracts. Dive deeper into the inner workings of the protocol.
+        Look up Rarimo APIs, JavaScript libraries, and smart contracts. Dive
+        deeper into the inner workings of the protocol.
       </>
     ),
     url: "./reference",